backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pulp
pydantic
openpyxl
python-multipart
